+++ packages/expo-router/src/layouts/Stack.web.tsx
@@ -1,11 +1,7 @@
+// This will be changed to `ExperimentalModalStack` in @expo/cli/src/start/server/metro/withMetroMultiPlatform.ts
+// When the `EXPO_UNSTABLE_WEB_MODAL` env variable is truthy.
+import Stack from './_web-modal';
 import { Screen } from '../views/Screen';
-
-const Stack = (() => {
-  if (process.env.EXPO_PUBLIC_EXPERIMENTAL_WEB_MODAL === '1') {
-    return require('./ExperimentalModalStack').default;
-  }
-  return require('./BaseStack').default;
-})();
 
 Stack.Screen = Screen;